// src/string-formatter-options.type.ts
/**
 * Options to configure how {@link StringFormatter} behaves.
 *
 * @version 1.0.0
 * @author Giancarlo Dalle Mole
 * @since 10/02/2020
 */
export type StringFormatterOptions = {

    /**
     * @default false
     */
    reverse?: boolean;
    /**
     * @default false
     */
    useDefaults?: boolean;
}